feat: Separate SLOs for node level monitoring in a service
What
-------------------------------------------------------------------
Provisions separately defining node-level SLO monitoring thresholds.

Why
-------------------------------------------------------------------
Services with high probability of creating hotspots on a single VM/instance
e.g. gitaly, need to have a node level SLO with greater tolerance for alerts
because a single user or project can get affected but overall service status
or a vast majority of the users will not get affected. We already have a provision
to create `SingleNode` alerts for a service but currently the generated alerts
use same threshold as the whole service and is [proven to be high in case of Gitaly](https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23576#note_1379536934). This has caused a string of
alerts for EOCs which were auto-resolved and didn't require EOCs intervention.
Separating threshold for a single node will allow us to lower the threshold for a single node.

Reference Issue: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/23576